feat: gas optimizations (#17)

+++ src/IguildCouncil.sol
@@ -3,13 +3,13 @@ pragma solidity ^0.8.9;
 
 interface GuildCouncilI {
 
-    function _guildVerdict(bool guildAgreement, uint256 proposalId)
+    function _guildVerdict(bool guildAgreement, uint48 proposalId)
         external
         returns(bool success);
-    function _callGuildsToVote(uint256[] calldata guildsId, uint256 proposalId)
+    function _callGuildsToVote(uint256[] calldata guildsId, uint48 proposalId)
        external
        returns(bool);
-    function sendSilver(address sender, address receiver, uint256 guildId, uint256 silverAmount)
+    function sendSilver(address sender, address receiver, uint48 guildId, uint256 silverAmount)
         external;
 
 }
